Extract `FeedStoreSpy` to shared helper
