+++ EssentialFeedTests/Feed Cache/CacheFeedUseCaseTests.swift
@@ -178,37 +178,3 @@ extension CacheFeedUseCaseTests {
         return NSError(domain: "any error", code: 1, userInfo: nil)
     }
 }
-
-
-// MARK: FeedStoreSpy Test Case Class
-extension CacheFeedUseCaseTests {
-    
-    private class FeedStoreSpy: FeedStore {
-        typealias OperationCompletion = (Error?) -> Void
-        
-        enum Message: Equatable {
-            case deleteCachedFeed
-            case insert([LocalFeedImage], Date)
-        }
-        
-        private(set) var receivedOperations = [(operation: Message, completion: OperationCompletion)]()
-        
-        func deleteCachedFeed(completion: @escaping OperationCompletion) {
-            receivedOperations.append((.deleteCachedFeed, completion))
-        }
-        
-        func completeDeletion(error: Error? = nil, at index: Int = 0) {
-            guard receivedOperations[index].operation == .deleteCachedFeed else { return }
-            receivedOperations[index].completion(error)
-        }
-        
-        func insert(_ feed: [LocalFeedImage], _ timestamp: Date, completion: @escaping OperationCompletion) {
-            receivedOperations.append((.insert(feed, timestamp), completion))
-        }
-        
-        func completeInsertion(error: Error? = nil, at index: Int = 1) {
-            guard receivedOperations[index].operation != .deleteCachedFeed else { return }
-            receivedOperations[index].completion(error)
-        }
-    }
-}
