+++ EssentialFeedTests/Feed Cache/LoadCachedFeedUseCaseTests.swift
@@ -18,7 +18,6 @@ class LoadCachedFeedUseCaseTests: XCTestCase {
     }
 }
 
-// MARK: FeedStoreSpy Test Case Class
 extension LoadCachedFeedUseCaseTests {
     private func makeSUT(currentDate: @escaping () -> Date = Date.init,
                          file: StaticString = #filePath,
@@ -29,34 +28,5 @@ extension LoadCachedFeedUseCaseTests {
         assertNoMemoryLeaks(sut, objectName: "LocalFeedLoader_Cache", file: file, line: line)
         
         return (sut, store)
-    }
-    
-    private class FeedStoreSpy: FeedStore {
-        typealias OperationCompletion = (Error?) -> Void
-        
-        enum Message: Equatable {
-            case deleteCachedFeed
-            case insert([LocalFeedImage], Date)
-        }
-        
-        private(set) var receivedOperations = [(operation: Message, completion: OperationCompletion)]()
-        
-        func deleteCachedFeed(completion: @escaping OperationCompletion) {
-            receivedOperations.append((.deleteCachedFeed, completion))
-        }
-        
-        func completeDeletion(error: Error? = nil, at index: Int = 0) {
-            guard receivedOperations[index].operation == .deleteCachedFeed else { return }
-            receivedOperations[index].completion(error)
-        }
-        
-        func insert(_ feed: [LocalFeedImage], _ timestamp: Date, completion: @escaping OperationCompletion) {
-            receivedOperations.append((.insert(feed, timestamp), completion))
-        }
-        
-        func completeInsertion(error: Error? = nil, at index: Int = 1) {
-            guard receivedOperations[index].operation != .deleteCachedFeed else { return }
-            receivedOperations[index].completion(error)
-        }
     }
 }
